Ignore presentation time if visibility changed, for Event Timing
EventTiming registers for presentation feedback times for events which
are expected to have a visual update.  However, if page visibility
changes, that can affect the time it takes to present the next paint
update.

We were already ignoring the time to next paint portion for UKM
reporting of interactions, but now we also ignore this time for Event
Timing reporting to Web Performance timeline.

Filed a spec issue with EventTiming API to clarify cases where events
will not have a next paint:
https://github.com/w3c/event-timing/issues/123

Also, we will follow-up to see if it is not desirable to measure all the
way until visibility change, as per:
https://github.com/w3c/event-timing/issues/129

Bug: 1312568
Change-Id: Iee0b5eee81fb216a96ecbfcc41ca26601ab6ba1b