[Sub Apps] Make APIs operate with root-relative URL paths
Design doc go/multi-apps

Currently the APIs (Add/List/Remove) operate with full URLs, change them
so they operate on root-relative URLs ("absolute path").

This makes them a bit more ergonomic to use for developers, and it
completely removes a class of errors - arguments having a different
origin to that of the calling app.

Also fixed an error where in `SubAppsServiceImpl::List` we were checking
whether the calling app was installed before the registry was
potentially ready.

Also some minor cleanups mostly in the browsertest.

Test: out/cros/browser_tests --gtest_filter=SubAppsServiceImplBrowserTest.\*

Bug: 1366301
Change-Id: I48ba1dbadc34ddcfcc256fd92d2f810a9861dbd8