[@scope] Support implicit scoping roots
This CL makes it possible to specify a @scope rule without any
prelude, scoping the rules implicitly to the parent of the owner node.

This is implemented by storing a StyleSheetContents in each StyleScope,
and treating each owner-parent-node of the StyleSheetContents as a
scoping root.

The following likely doesn't work, and is postponed for future CLs:

 - Constructed stylesheets
 - An owner node whose parent node is a ShadowRoot

https://github.com/w3c/csswg-drafts/issues/7349

Bug: 1379844
Change-Id: Icbfae88636662e9a16dd11f9c7a371c997447f60