Add setMetadata() support for VP8.
This covers everything in the WebIDL except for the RTP related
metadata because that is not (yet?) part of the
webrtc::VideoFrameMetadata. Let's focus on one thing at a time and deal
with that separately.

For now, the WPT still passes because we're using clone() instead of
an empty constructor. Today the clone() method takes care of copying
the RTP related metadata for us, but it does not take care of the
encoder specifics.

In a separate CL we should refactor clone() as
"constructor + set bytes + setMetadata" so that we don't get any free
rides from it. The goal is that "what you see is what you get".

Bug: webrtc:14709
Change-Id: I9da91a68cd6c7bd2b3a8a5a4ac62d6b70a95858c