[Web Platform Test] Just-in-time install for change-payment-method
Before this patch, the change-payment-method-manual.https.html Web
Platform Test (WPT) was installing a payment handler using the
PaymentInstruments API, which is planned to be removed.

This patch adds a payment method manifest and a web app manifest (in the
same file) for the payment handler under test, which allows it to be
installed just-in-time (JIT) when PaymentRequest.show() is called.

After this patch, the change-payment-method-manual.https.html WPT no
longer uses the PaymentInstruments API for manual payment handler
installation.

Bug: 1327265
Change-Id: Ie399d77a943167aa5a4ffaba034f3b6a00a068c2